+++ src/games/tank/scenes/menu-scene.ts
@@ -20,46 +20,33 @@ export class MenuScene extends Phaser.Scene {
       Phaser.Input.Keyboard.KeyCodes.S
     );
     this.startKey.isDown = false;
-    this.initRegistry();
   }
 
   create(): void {
     this.bitmapTexts.push(
       this.add.bitmapText(
-        this.sys.canvas.width / 2 - 65,
+        this.sys.canvas.width / 2 - 120,
         this.sys.canvas.height / 2,
         "font",
         "PRESS S TO PLAY",
-        8
+        30
       )
     );
 
     this.bitmapTexts.push(
       this.add.bitmapText(
-        this.sys.canvas.width / 2 - 60,
-        this.sys.canvas.height / 2 - 40,
+        this.sys.canvas.width / 2 - 120,
+        this.sys.canvas.height / 2 - 100,
         "font",
-        "SPACE INVADERS",
-        8
+        "TANK",
+        100
       )
     );
   }
 
   update(): void {
     if (this.startKey.isDown) {
-      this.scene.start("HUDScene");
       this.scene.start("GameScene");
-      this.scene.bringToTop("HUDScene");
     }
-  }
-
-  /**
-   * Build-in global game data manager to exchange data between scenes.
-   * Here we initialize our variables with a key.
-   */
-  private initRegistry(): void {
-    this.registry.set("points", 0);
-    this.registry.set("lives", 3);
-    this.registry.set("level", 1);
   }
 }